Renamed containers that function as pages to include page in the name
