+++ client/src/App.js
@@ -7,17 +7,17 @@ import { setInitialAuth } from './redux/actions/auth';
 import * as s from './styles';
 import Landing from './components/Presentational/Landing';
 import Logout from './components/Logout';
-import Login from './components/Container/Login';
+import Login from './components/Container/LoginPage';
 import Register from './components/Register';
 import RegistrationSuccess from './components/RegistrationSuccess';
-import LoginEmployee from './components/Container/LoginEmployee';
-import CreateEmployee from './components/Container/CreateEmployee';
+import LoginEmployee from './components/Container/LoginEmployeePage';
+import CreateEmployee from './components/Container/CreateEmployeePage';
 import Servers from './components/Servers';
 import TablesPage from './components/Container/TablesPage';
 import NotFound from './components/Presentational/NotFound';
 import Navbar from './components/Presentational/Navbar';
 import PartyPage from './components/Container/PartyPage';
-import Settings from './components/Container/Settings';
+import Settings from './components/Container/SettingsPage';
 import Sidebar from './components/Presentational/Sidebar';
 import NewRestaurant from './components/NewRestaurant';
 import RequireNotAuth from './components/HOC/RequireNotAuth';
